requirements.txt
behave==1.2.6
flake8==3.9.2
selenium==4.1.0
sure==2.0.0
webdriver-manager==3.5.2
mypy==0.910
mypy-extensions==0.4.3
typing-extensions==3.10.0.2
pre-commit==2.16.0
SQLAlchemy==1.4.31
python-dotenv==0.19.2
PyMySQL==1.0.2
